fix: Float literal type and parenthesized expression integration
- Change Value.Double(value) to Value.Float(value.toFloat()) in floatLiteral()
  parser to use correct Value.Float type per spec
- Add grouping() to primary() alternatives, tried before identifiers
  to enable parenthesized expressions like (1 + 2)

+++ ink/src/main/kotlin/org/inklang/peg/ink/expressions/Literals.kt
@@ -92,7 +92,7 @@ fun floatLiteral(): Parser<Expr> {
             val value = lexeme.toDoubleOrNull() ?: return ParseResult.Failure("float literal", position)
 
             return ParseResult.Success(
-                Expr.LiteralExpr(Value.Double(value)),
+                Expr.LiteralExpr(Value.Float(value.toFloat())),
                 end
             )
         }
